Rearranged App file for cleaner structure. Put navbar in Header so we can change the language if need be.
