+++ index.html
@@ -9,17 +9,6 @@
 </head>
 
 <body>
-    <div id="navbar">
-        <div id="navTitle">
-            LingView: ELAN and FLEx Web Display
-        </div>
-        <ul id="navLinks">
-             <li><a href="#/search/">Search</a></li>
-             <li><a href="#/about/">About</a></li>
-             <li><a href="#/glossary/">Glossary</a></li>
-             <li><a href="#/index/">Index of Texts</a></li>
-        </ul>
-    </div>
     <div id="main"></div>
     <script src="build/bundle.js"></script>
     <script>
